K3
  * Changed BW selector to include finer steps
  * Added read options to initialization method.
    - set power control iaw availability of K3/100 module
    - required a change in get_pc_min_max method - forced
      change in non K3 rig classes
  * Added default Mode/BW table
  * Changed Power out method to use the BG; query

+++ src/include/TS480HX.h
@@ -47,7 +47,7 @@ public:
 	int  get_volume_control();
 	void set_power_control(double val);
 	int  get_power_control();
-	void get_pc_min_max_step(int &min, int &max, int &step) {
+	void get_pc_min_max_step(double &min, double &max, double &step) {
 		min = 5; max = 200; step = 1; }
 	void set_attenuator(int val);
 	int  get_attenuator();
